+++ lib/loadrunner/server_helper.rb
