+++ lib/loadrunner/signature_helper.rb
