+++ loadrunner.gemspec
@@ -1,7 +1,7 @@
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require 'date'
-require 'load_runner/version'
+require 'loadrunner/version'
 
 Gem::Specification.new do |s|
   s.name        = 'loadrunner'
