+++ spec/loadrunner/client_spec.rb
