+++ spec/loadrunner/command_line_spec.rb
