+++ spec/loadrunner/signature_helper_spec.rb
